edit

+++ printf/ft_print_string.c
@@ -6,7 +6,7 @@
 /*   By: hkubo <hkubo@student.42tokyo.jp>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/03/13 17:44:52 by hkubo             #+#    #+#             */
-/*   Updated: 2021/03/13 21:24:02 by hkubo            ###   ########.fr       */
+/*   Updated: 2021/03/13 21:24:28 by hkubo            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -51,7 +51,7 @@ int field_no_flag(char *str, a_list flag_list)
     int len;
     
     len = 0;
-    if (flag_list.precision != -1 && flag_list.precision < ft_strlen(str))
+    if (flag_list.precision != -1 && flag_list.precision < (int)ft_strlen(str))
     {
         i = flag_list.precision;
         while (flag_list.field - i)
